Initial README
This is the initial README that is based on the converation I had in the HipChat Training channel one night. This is not word-for-word and is some cases misses some of the sense of adventure and story that I wrote about. But this is the distilled version that hopefully allow others to contribute.