Adds typing out of scene and option descriptions and current act
To provide some sanity to the layout of a game I thought it would be better
if there was an optional way to describe the act that it belong to.
Seeings as 'another-level' is main level I thought it could go to one of the
sub-levels when complete.

The current act part is working.
The typing out is working.

Signed-off-by: Franklin Webber <franklin@chef.io>