Ruby based scene dsl
Instead of advancing the story I'm working on the scene structure.

The lofty goal I'm after now is to set the correct scene when I transition to the command-line. Because I think I can have the scenes continue for the player so it's as if they never really left the interface.

Signed-off-by: Franklin Webber <franklin@chef.io>