Adds tree the game world and conditions around that
Testing out how multiple conditions behave. The slowness of unloading and loading ruby a bunch is beginning to take its toll on the speed of the prompt leaving the game and then essentially running the game code again. The story of the woodpecker and the oak is taking a new direction. Instead of trying to push in some Chef concepts I am simply having them do some linux commands. Which might be the answer for an interactive experience for the developer essentials material. Buried in its bowels. It would be a safe starter and we're only about two commands away from convering all that it covers.

Signed-off-by: Franklin Webber <franklin@chef.io>