Updates to the tree, speech, and copy
The names of the branches of the tree were confusing with the 'a' article.
I updated the copy for a lot of the scenes and brought the colored text back to
the introduction.
I'm seeing some small issues with the speech stroke break across lines
incorrectly but leaving it for now.
The unique, random id generated for the options does not work becaue the ID for each option is generated roughly at the same time (within the second).
Gave the player a choice to remove / move the termites.
The termites have a joke with a clearer punch line.

Signed-off-by: Franklin Webber <franklin@chef.io>