Completed initial verison - Woodpecker and the Oak
This is the story of a woodpecker and an oak tree. The focus of the learning
is on basic linux skills of working with the file system. It still needs more
polish in that direction.

This works through the majority of the issues and limitations to get the
story completed.

Signed-off-by: Franklin Webber <franklin@chef.io>